chore: update user management methods

+++ wetalkAPI/wetalkAPI/Controllers/UsersController.cs
@@ -35,39 +35,6 @@ namespace MovieseekAPI.Controllers
         }
 
         /// <summary>
-        /// Gets all users
-        /// </summary>
-        /// <response code="201">Returns the users</response>
-        /// <response code="401">Unauthorized</response>
-        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
-        [HttpGet]
-        public IActionResult GetAll()
-        {
-            var users = _userService.GetAll();
-            var model = _mapper.Map<IList<UserModel>>(users);
-            return Ok(model);
-        }
-
-        /// <summary>
-        /// Get a user by id
-        /// </summary>
-        /// <response code="200">Returns the user</response>
-        /// <response code="401">Unauthorized</response>            
-        /// <response code="404">User wasn't found</response>
-        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
-        [HttpGet("{id}")]
-        public IActionResult GetById(int id)
-        {
-            var user = _userService.GetById(id);
-            if (user != null)
-            {
-                var model = _mapper.Map<UserModel>(user);
-                return Ok(model);
-            }
-            return NotFound();
-        }
-
-        /// <summary>
         /// Authenticate a user
         /// </summary>
         /// <response code="200">Returns the user with access token</response>
@@ -78,7 +45,7 @@ namespace MovieseekAPI.Controllers
         {
             var user = _userService.Authenticate(model.Username, model.Password);
 
-            if (user == null)
+            if (user == null || (user != null && user.Active == 0))
                 return BadRequest(new { message = "Username or password is incorrect" });
 
             var tokenHandler = new JwtSecurityTokenHandler();
@@ -128,6 +95,7 @@ namespace MovieseekAPI.Controllers
             {
                 // create user
                 _userService.Create(user, model.Password);
+
                 AuthenticateUserModel authModel = new AuthenticateUserModel()
                 {
                     Username = user.Username,
@@ -143,12 +111,129 @@ namespace MovieseekAPI.Controllers
         }
 
         /// <summary>
+        /// Gets all users
+        /// </summary>
+        /// <response code="201">Returns the users</response>
+        /// <response code="401">Unauthorized</response>
+        /// <response code="403">Forbidden</response>
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
+        [HttpGet]
+        public IActionResult GetAll()
+        {
+            if (User.Identity.Name != null)
+            {
+                // get user permissions
+                var currentUser = _userService.GetById(int.Parse(User.Identity.Name));
+
+                if (currentUser == null)
+                {
+                    BadRequest("Current user not found");
+                }
+
+                if (currentUser.PermissionID == 1)
+                {
+                    var users = _userService.GetAll();
+                    var model = _mapper.Map<IList<UserModel>>(users);
+                    return Ok(model);
+                }
+            }
+
+            return Forbid();
+        }
+
+        /// <summary>
+        /// Get a user by id
+        /// </summary>
+        /// <response code="200">Returns the user</response>
+        /// <response code="401">Unauthorized</response>            
+        /// <response code="403">Forbidden</response>
+        /// <response code="404">User wasn't found</response>
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
+        [HttpGet("{id}")]
+        public IActionResult GetById(int id)
+        {
+            if (User.Identity.Name != null)
+            {
+                // get user permissions
+                var currentUser = _userService.GetById(int.Parse(User.Identity.Name));
+
+                if (currentUser == null)
+                {
+                    BadRequest("Current user not found");
+                }
+
+                if (currentUser.PermissionID == 1)
+                {
+                    var user = _userService.GetById(id);
+                    if (user != null)
+                    {
+                        var model = _mapper.Map<UserModel>(user);
+                        return Ok(model);
+                    }
+                    return NotFound();
+                }
+            }
+
+            return Forbid();
+        }
+
+        /// <summary>
+        /// Creates a user
+        /// </summary>
+        /// <response code="200">Returns the created user</response>
+        /// <response code="400">Bad request</response>            
+        /// <response code="403">Forbidden</response>            
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
+        [HttpPost]
+        public IActionResult Create([FromBody] CreateUserModel model)
+        {
+            if (!ModelState.IsValid)
+            {
+                ModelState.AddModelError("error", "invalid request body");
+                return BadRequest(ModelState);
+            }
+
+            // map model to entity
+            var user = _mapper.Map<User>(model);
+
+            try
+            {
+                if (User.Identity.Name != null)
+                {
+                    // get user permissions
+                    var currentUser = _userService.GetById(int.Parse(User.Identity.Name));
+
+                    if (currentUser == null)
+                    {
+                        BadRequest("Current user not found");
+                    }
+
+                    if (currentUser.PermissionID == 1)
+                    {
+                        // create user
+                        _userService.Create(user, model.Password);
+
+                        var result = _mapper.Map<UserModel>(user);
+                        return Ok(result);
+                    }
+                }
+
+                return Forbid();
+            }
+            catch (Exception ex)
+            {
+                // return error message if there was an exception
+                return BadRequest(new { message = ex.Message });
+            }
+        }
+
+        /// <summary>
         /// Updates an user
         /// </summary>
         /// <response code="200">Returns the updated user</response>
         /// <response code="400">Bad request</response>            
         /// <response code="401">Unauthorized</response>            
-        /// <response code="404">Movie wasn't found</response>
+        /// <response code="404">User wasn't found</response>
         [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         [HttpPut("{id}")]
         public IActionResult Update(int id, [FromBody] UpdateUserModel model)
@@ -165,11 +250,27 @@ namespace MovieseekAPI.Controllers
 
             try
             {
-                // update user 
-                _userService.Update(user, model.Password);
-                var updatedUser = _userService.GetById(id);
-                var mappedUser = _mapper.Map<UserModel>(updatedUser);
-                return Ok(mappedUser);
+                if (User.Identity.Name != null)
+                {
+                    // get user permissions
+                    var currentUser = _userService.GetById(int.Parse(User.Identity.Name));
+
+                    if (currentUser == null)
+                    {
+                        BadRequest("Current user not found");
+                    }
+
+                    if (currentUser.PermissionID == 1)
+                    {
+                        // update user
+                        _userService.Update(user, model.Password);
+                        var updatedUser = _userService.GetById(id);
+                        var mappedUser = _mapper.Map<UserModel>(updatedUser);
+                        return Ok(mappedUser);
+                    }
+                }
+
+                return Forbid();
             }
             catch (Exception ex)
             {
@@ -188,13 +289,40 @@ namespace MovieseekAPI.Controllers
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
-            var user = _userService.GetById(id);
-            if (user != null)
+            try
             {
-                _userService.Delete(id);
-                return NoContent();
+                if (User.Identity.Name != null)
+                {
+                    // get user permissions
+                    var currentUser = _userService.GetById(int.Parse(User.Identity.Name));
+
+                    if (currentUser == null)
+                    {
+                        BadRequest("Current user not found");
+                    }
+
+                    if (currentUser.PermissionID == 1)
+                    {
+                        var user = _userService.GetById(id);
+                        if (user != null)
+                        {
+                            _userService.Delete(id);
+                            return NoContent();
+                        }
+                        return NotFound();
+                    }
+                }
+
+                return Forbid();
             }
-            return NotFound();
+            catch (Exception ex)
+            {
+                // return error message if there was an exception
+                return BadRequest(new
+                {
+                    message = ex.Message
+                });
+            }
         }
     }
 }
